.NET runtime: Fan.Sys.Int => Fan.Sys.Long

+++ src/nfan/fan/sys/FuncType.cs
@@ -37,7 +37,7 @@ namespace Fan.Sys
   // Type
   //////////////////////////////////////////////////////////////////////////
 
-    public override Int hash()
+    public override Long hash()
     {
       return signature().hash();
     }
